Translations
Added option to add Translations
Fixed (again) ESL 100+ not having proper index in BOS file

+++ src/Utils.cpp
@@ -123,7 +123,7 @@ namespace Utils {
         if (!dataHandler) return nullptr;
 
         if (modIndex == 0xFE) {
-            uint8_t eslIndex = (formID >> 12) & 0xFFF;
+            uint16_t eslIndex = (formID >> 12) & 0xFFF;
             return dataHandler->LookupLoadedLightModByIndex(eslIndex);
         }
